Add code and tests to help with timestamp handling
It was not intuitive to me how to interact with timestamp columns via
this library so I've added some code.

I think it might be reasonable to have some kind of
pgo/dynamic.timestamp or something to help with decoding the values
instead of having the nested tuples and ints. There might be other
approaches though.

I'm still new to all this and just trying to contribute but I'm
completely happy if this isn't deemed appropriate for this library
for some reason.