Add a 'decode_timestamp' function
To help with decoding the nested tuple format which is how erlang's
pgo handles timestamps. Named as suggested by lpil.

I'm not sure the calling convention for the functions and so whether
or not it should be set up to use parens at the call site. I've opted
not to, to follow the convention of other value types like 'int' and
'string'.